+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 3.5.0 (IN PROGRESS)
+
+### Features / Enhancements
+
+- Add disable if code (#321)
+- Update reset request visibility if reset button is hidden (#322)
+- Add code options source for select element (#323)
+
 ## 3.4.0 (2023-12-14)
 
 ### Features / Enhancements